BRICK-72-feat: mockImages.ts 생성
