+++ components/ImageViewer/ImageViewer.tsx
@@ -5,37 +5,7 @@ import Image from 'next/image'
 import usePortal from '@/hooks/usePortal'
 import { ImageViewerModal, Portal } from '@/components'
 import moreImgIcon from '@/public/images/svgs/moreImg.svg'
-
-const mockImages = [
-  {
-    id: 1,
-    src: 'https://cdn.pixabay.com/photo/2024/02/15/15/46/cat-8575641_1280.jpg',
-  },
-  {
-    id: 2,
-    src: 'https://cdn.pixabay.com/photo/2023/12/04/16/12/leaf-8429779_1280.jpg',
-  },
-  {
-    id: 3,
-    src: 'https://cdn.pixabay.com/photo/2024/01/02/13/28/gecko-8483282_1280.png',
-  },
-  {
-    id: 4,
-    src: 'https://cdn.pixabay.com/photo/2024/03/28/15/59/life-8661191_1280.jpg',
-  },
-  {
-    id: 5,
-    src: 'https://cdn.pixabay.com/photo/2023/12/04/16/12/leaf-8429779_1280.jpg',
-  },
-  {
-    id: 6,
-    src: 'https://cdn.pixabay.com/photo/2024/01/02/13/28/gecko-8483282_1280.png',
-  },
-  {
-    id: 7,
-    src: 'https://cdn.pixabay.com/photo/2023/12/04/16/12/leaf-8429779_1280.jpg',
-  },
-]
+import mockImages from './mockImages'
 
 function ImageViewer() {
   const { portalRef, isPortalOpen, setIsPortalOpen, handleOutsideClick } =
